Prevent App::Satpass2->run() from prompting after an end-of-file. The fix was in two parts:
Make the execute() method return after a single command if the first
argument was a code reference. The presumption is that the code
reference knows where it is in the input and will pick up where it left
off.

Make _get_readline() return a code reference that knows when it has hit
end-of-file, and return undef thereafter. This would suffice for the
fix, but was necessary only for the case where the end-of-file was in
response to a continuation prompt. The fix to execute() was done anyway
because it made the code work more like it was intended to work.


git-svn-id: file:///Users/tom/Code/svn/repos/App-Satpass2/trunk@1846 f6fd3633-1913-47c2-bad4-372d75c948ec