This commit is a bit of a mish-mash.
I decided to disallow ${1:=foo}, for consistency with Bash. This meant
recoding the test. I used an environment variable for this, which meant
if it was undefined it got exported, which caused further failures until
I invented an unexport() method. Documented all this.

While I was in the documentation, I documented the use of environment
variable SATPASSINI as part of a last-ditch effort to find an
initialization file.


git-svn-id: file:///Users/tom/Code/svn/repos/App-Satpass2/trunk@1897 f6fd3633-1913-47c2-bad4-372d75c948ec