Add iridium.dat to the top-level directory's svn:ignore.
Add optional templates pass_horizon, pass_maximum, pass_illumination,
and pass_appulse to App::Satpass2::Format::Classic, to override the
'pass' template for the appropriate event. The latter replaces the old
template of the same name, and defaults to the 'pass' template plus the
information in the old 'pass_appulse' template. The other new templates
default to undef, which causes the 'pass' template to be used.


git-svn-id: file:///Users/tom/Code/svn/repos/App-Satpass2/trunk@1997 f6fd3633-1913-47c2-bad4-372d75c948ec