Tweak the code in App::Satpass2->dispatch(), with the thought of centralizing the call to the dispatched code, in case I want to move the interactive flag to the call frame.
Tweak docs, including documenting that macros can be called
programmatically via the dispatch() method. I actually got callable
macros working via AUTOLOAD, but was not happy with the code fragment
returned by can() (it had to be called on the object that generated it,
but there was no easy way to ensure that happened -- also it became
invalid if the macro was deleted).


git-svn-id: file:///Users/tom/Code/svn/repos/App-Satpass2/trunk@2014 f6fd3633-1913-47c2-bad4-372d75c948ec