Create Astro::App::Satpass2::Utils to hold utility subroutines. This is private to the package.
Move __instance and __quoter from Astro::App::Satpass2::Copier to
Astro::App::Satpass2::Utils, losing the leading '__' and documenting
them in the process.

Create utility method load(), which takes a module name and optional
prefixes to try, and attempts to load the module, using a stringy eval.
If the module can be loaded, the full module name (with prefix if any)
is returned. If not, undef is returned. Results are cached, so that
subsequent calls with the same arguments return whatever is in the
cache.

Use the new load() subroutine in Astro::App::Satpass2::Format, to load
the time_formatter class if needed, and with the prefix
'Astro::App::Satpass2::FormatTime::'.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@2140 f6fd3633-1913-47c2-bad4-372d75c948ec