Have Astro::App::Satpass2::FormatValue set the time of the Astro::Coord::ECI objects on the first formatter call rather than on creation.
This lets the events() method return a reference to a list of
Astro::App::Satpass2::FormatValue objects, rather than an iterator
function that creates each one as it is used.

This in turn lets me rewrite the 'pass' canned template in terms of
FOREACH rather than WHILE.

And this (and here's the point of the whole thing) lets me get rid of
the logic to fiddle with the Template-Toolkit WHILE_MAX variable. This
is a win because you were so restricted on when you could manupulate
that variable if you wanted the change to be seen by the templating
system.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@2270 f6fd3633-1913-47c2-bad4-372d75c948ec