In inc/, make the Astro::App::Satpass2::Test::App method() subroutine take a code reference as its $want argument. This is handled by loading @_ with ( $want, $got, $title ) and executing a goto to it.
Provided exports TRUE, FALSE, and INSTANTIATE from
Astro::App::Satpass2::Test::App. TRUE and FALSE test if the result of
the operation was true or false respectively. INSTANTIATE is like TRUE,
but in addition it stores $got as the object to be further tested.

Modified all tests to use TRUE, FALSE, and INSTANTIATE where
appropriate.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@2343 f6fd3633-1913-47c2-bad4-372d75c948ec