Get rid of the individual formatting methods in the Astro::App::Satpass2::Format subclasses, and instead have one format() method, which takes two arguments: the name of the format and the data (usually an array or hash reference). Obviously the relevant Astro::App::Satpass2 methods and test scripts need to change to track this.
In Astro::App::Satpass2::Format::Template, format() needs to be smart
enough to call _wrap() when needed (which is every time except for
'alias'). To simplify the job of Astro::App::Satpass2::FormatValue a
bit, the Astro::App::Satpass2 phase() method passes a hash with {body}
and {time} keys, rather than simply a body set to the correct time.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@2447 f6fd3633-1913-47c2-bad4-372d75c948ec