Modify Astro::App::Satpass2 to allow file names, URLs, open file handles, scalar references, list references, or code references as input to the init(), load(), and source() methods. The revised functionality all goes through the _file_reader() method, which ordinarily (but see below) returns a code snippet which returns a line of input every time it is called.
The source() method actually allowed most of these (not URLs or code
references), but although init() was written as a wrapper for source(),
it disallowed anything but files. The change generifies the code and
uses the same code in all three locations.

Well, almost the same. It did not make sense to me to take a scalar
reference to HTTP::Result->content(), turn it into a file handle, then
glob that filehandle for the load() method. So if you want the whole
thing at once, _file_reader() takes as its second argument an options
hash. This is normally used to implement the -optional option of
source(), but support for { glob => 1 } was added, to have the contents
of the file returned.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@2805 f6fd3633-1913-47c2-bad4-372d75c948ec