Add a title_gravity mutator to Astro::App::Satpass2::FormatValue to control whether multiline titles are aligned to the top or the bottom of their space.
Aligning to the bottom involves generating an extra blank line at the
top of the titles, so that we get one pass through all of them to find
out how many lines we need, before we start putting title text onto the
template. This can be dealt with by fiddling with the template in cases
where there is non-title text before the titles (e.g. the default
'position' template). For those where the titles come first, we remove
the first line of the template output _IF_ it is empty, and _IF_ the
title object, after the template is executed, is set to align to the
bottom.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@2817 f6fd3633-1913-47c2-bad4-372d75c948ec