This commit is a hodge-podge.
It started off as an attempt to have the geocoding wrappers load the
wrapped modules at run time, so that various Kwalitee tools do not
believe that there is an undeclared dependency on the wrapped modules.

That required Astro::App::Satpass2 to be tweaked, because now the
wrappers could be loaded even when the wrapped modules were unavailable,
which caused Astro::App::Satpass2 to choose an unusable default
geocoder.

While I was at it, I eliminated unused 'use Carp;' from the geocoding
wrappers. I also eliminated it from Astro::App::Satpass2::Warner, and
instead had it 'require Carp;' where needed. The intent is not to 'use
Carp;' at all, to try to get a faster load.

The idea behind Astro::App::Satpass2::Warner was that it could be
instantiated once and passed around, so that everyone did the same style
of warning. But it turned out the passing around was not being done,
because while Astro::App::Satpass2 was routing all the relevant
instantiation through its _set_copyable() method, that method was not
passing the warner into any of the objects it handled.

So I added 'warner => $self->{_warner},' to _set_copyable(), only to
discover that the Astro::App::Satpass2::ParseTime objects were not using
named arguments to new(), but instead were just doing a list of possible
classes to try. So that got converted to take a 'class => $string'
argument to new(), where the $string is a comma-delimited list of valid
class names.

Then all the Astro::App::Satpass2::ParseTime tests had to be rewritten
to use the new syntax for specifying that subclasses be tried in a given
order.

While I was mucking around with loading modules,
Astro::App::Satpass2::Util::load_package() grew some validation on
package names.

And Astro::App::Satpass2::Copier got an init() method, which takes
name/value pairs and sets the corresponding attributes. It dies if given
a name that does not represent a known attribute. The ParseTime,
Formatter, and Geocoder new() methods call this with any arguments they
have not already processed.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@3024 f6fd3633-1913-47c2-bad4-372d75c948ec