The t/whole_app.t 'cd to home' test persists in failing for exactly one tester. The failure is that when I cd to "/home/foo" Cwd::cwd reports that I am in "/usr/home/foo".
The failure occurs under FreeBSD 6.2, which was released January 2007,
and for which I can no longer find any installation CDs on line. I could
load from floppies, but I am not very much inclined to try, since all
the other FreeBSD testers are at 9.0, and since that failed to install
under Oracle VirtualBox.

So I'm attempting to trap the failing case, turn the test into a TODO,
and dump some possibly-useful data (the contents of $ENV{HOME} and the
result of Cwd::getcwd).

I have also put a paragraph in the Astro::App::Satpass2 cd()
documentation about this.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@3058 f6fd3633-1913-47c2-bad4-372d75c948ec