Fixed failure in Astro::App::Satpass2::Format->decode( 'desired_equinox_dynamical' ) when the value was not 0.
Made ~~/... expand to configuration directory

Reimplemented macros as classes. The traditional macros are implemented
as Astro::App::Satpass2::Macro::Command.

Added the ability to implement a macro as Perl code. These are handled
by Astro::App::Satpass2::Macro::Code. The corresponding satpass2 syntax
is 'macro load Module_name'.

Updated version of Astro::SpaceTrack needed to one that supports the
REST interface.

Require Astro::Coord::ECI 0.57 because of the international launch
designator problems introduced in 0.051_01.

In the Astro::App::Satpass2 sky() method, change the use of
load_module() (from Astro::Coord::ECI::Utils) to load_package (which is
our own version of the functionality).

Make load_package() optionally callable as a method so it can use the
same error reporting mechanism as everyone else if it is available.

Add options to the load_package() calling sequence: lib (which directory
to add to @INC), complain (how to report validation failures), and fatal
(whether and how to report failures to load the module). A true value of
'fatal' causes the failure not to be cached so that another load of the
module will produce the same failure.

Made $satpass2->time_parser( base => ... ) both parse an input time and
return a formatted time. This is a preliminary to maybe changing whether
the base for a relative time is reset for each command.

Moved the __parse_time_reset() code to the __arguments() method, and
eliminated the calls to it in individual interactive methods and code
macros. The __parse_time_reset() method now returns nothing (it used to
return the base time but that was unused) and it simply returns if the
time_parser object is not set up, since it gets called while
instantiating Astro::App::Satpass2.

Made time parsing conform to the documentation in that if the time was
defaulted, the default sets the base time for the parser.

Routed all accesses to the base time in Astro::App::Satpass2::ParseTime
through base(), in case any subclass plays games with it.


git-svn-id: file:///Users/tom/Code/svn/repos/Astro-App-Satpass2/trunk@4064 f6fd3633-1913-47c2-bad4-372d75c948ec