Build time formats when needed. Validate fmtrs.
Build time formats for individual formatters out of the date_format and
time_format attributes when they are needed, instead of up front. The
former way of doing things made it impossible for subclasses to add time
fields.

Have the __make_formatter_method() validate the formatter data. This
is desirable for subclasses.