Support extra events, quarter select, time round
This turned out to be a bit of an omnibus commit. The changes are:

The pass() method now supports body-specific events such as are provided
by the still-in-progress NASA KMZ file object.

The quarters() method got extra options -new, -first, -full, -last to
select those quarters.

Rounding of time for display became configurable. This turned out to be
a bit of a mess, because support had to go all over the place. The
Format object needed an attribute, the Format::Template object had to
pass it to FormatValue, which had to support it and pass it on to
FormatTime, and the various FormatTime objects had to actually do the
rounding as required.