Add formatters on a per-FormatValue basis.
This represents a way to define a formatter without subclassing
Astro::App::Satpass2::FormatValue. Instead, the formatter is named and
added to a hash, and dispatched using AUTOLOAD.

This still needs decent documentation, but before that I need to decide
how the titles (currently in the %locale hash for the built-ins) are
going to work.