Rationalize high-level formatter code.
This mostly means name the variables consistently so I can figure out
what is going on. But it also involved moving the formatter name into
its descriptor hash, so I don't need to pass it around as a separate
argument to the various subroutines. I have also simply passed around
the entire formatter description rather than pulling out pieces. This is
conceptually simpler, but also means if I put localization in the
individual descriptions, it is available to the title formatters.