Ground work for localizing messages & templates
Make __localize() (in A::A::S::Locale) return all localizations in
list context. This beccause I eventually need to find all defined
templates, when some (at least) of them are defiend in the localization.

Make sure all currently-coded calls to __localize() are called in scalar
context, forcing it with the scalar() built-in if necessary.

Add __message() to A::A::S::Locale

Have A::A::S::Warner call __message() in whinge(), wail(), and weep().