Try to kill DESTROY warnings in FormatValue.
The majority of smokers show these, though I am unable so far to
duplicate them. The problem is the use of AUTOLOAD to make user-defined
formatters work with Template-Toolkit. To be consistent with predefined
formatters they have to look like methods, but I can't define them in
the Astro::App::Satpass2::FormatValue name space because they are
object-specific.