Rewrite custom formatter code in FormatValue.
The previous implementation used AUTOLOAD(), but this caused problems in
most of the smokers due to DESTROY() not being defined. Rather than
figure out all the Perl-defined stuff I need to bypass or write stubs
for, I have gone to placing stub methods in the Astro::App::Satpass2
name space. The stubs dispatch to the object-specific formatter of the
same name, or throw an exception if the object does not have a formatter
by that name.