Allow perl -setup to save original arguments
This involved adding a subroutine attribute that tells the tokenizer
whether to do tilde expansion on the token list, and then using that to
turn off tilde expansion for the perl() method. I probably should never
have had the tokenizer do this, since it produces inconsistent behavior
between interactive and non-interactive method calls, but it's probably
too late to change now.