Skip t/parse_time_date_manip_6.t if zone problem
What I ran into was that Date::Manip was picking the wrong Olson zone,
but DateTime was getting it right, so I put in a cross check, which is
only done if DateTime::TimeZone can be loaded, and which skips if the
two sources come up with different zones.

The problem was that Date::Manip::TZ was trying to reconstruct the Olson
zone from 'BRT' (Brazilian time) and was "randomly" picking a location
that did not do summer time. This is mostly OK, since 'BRT' is winter
time ('BRST' is summer time), but if you feed it January 1 you get the
wrong time back.

Of course, I also skip if DateTime::TimeZone gets it wrong.