Misc. time handling changes.
Allow a scalar reference as a time. This is interpreted as epoch.

Document how epoch times are specified.

Redo DateTime formatting to allow (I hope) for other DateTime-ish
classes to be used. What I really want is DateTime::Calendar::Christian
to parallel the date display of (say) CalSky or Stellafane.

Special-case DateTime::Strftime format %{calendar_name} to display the
name of the calendar in use. For the moment this is always 'Gregorian'.

Yes, this should have been two or three commits.