Rewrite ParseTime::ISO8601
The problem was that pulling the zone off the end first could remove too
much. The case in point was 'year-month-day' (or 'year+month+day) which
interpreted the day as a zone offset.

So I left parsing the zone until last, which required deferring some
calculations until after the zone was known.

Also (and this should have been a separate commit) had the formatter
object display its time_formatter attribute as a short name where
possible.