Sanitize DateTime obj used to calc fld wid
This is a case of bugs beget more bugs. When I started passing a locale
to the instantiator of the DateTime object used to calculate field
widths, it broke becaue DateTime::Calendar::Christian did not accept
this argument. So I had to use ->now() instead, because
DateTime::Calendar::Christian just delegates to DateTime->now().

But that meant that on days near the end of the month, the feeling
around for the biggest time field could generate invalid dates (e.g. Feb
30 if run on the 30th). The fix was to do a bunch of set() calls to make
sure the object represented midnight January 1 of the current year.