Add testing method load_or_skip().
This replaces most of the 'eval { ... } or skip' code, but not all of it
because of the case where multiple modules have the same interface and
any of them can be used.

The incentive for this was to centralize conditional loading so that if
I wanted to add a version check I could do so.