Make if() shortcut 'and' and 'or'
As a consequence of this, 'attr spacetrack,username' (or other valid
attribute) is an error if Astro::SpaceTrack can not be loaded, because
if it can not be loaded we can not determine what attributes are valid.
Shortcuting lets us do this, because in the expression
'attr spacetrack and attr spacetrack.username', the right-hand operand
is not evaluated if the left-hand operand is false.

Also rationalized (IM(NS)HO) the infix engine code.