Modernize DateTime::Calendar::Christian code
The thing I could not fit in Git's 50 character limit was that there was
in Astro::App::Satpass2::FormatTime::DateTime a kluge that was put in
because DateTime::Calendar::Christian->new() did not support the
'locale' argument. Since it does as of version 0.06, and since this code
now requires 0.06, the kluge is now removed.