Handle inconsistency Time::Local vs Time::y2038
The problem here is that earlier in the year I found that Time::Local
worked better for dates well away from the current year if I passed it
the Gregorian year rather than computing the Perl year. But I support
Time::y2038 for the benefit of older Perls, and it expects a Perl year.
The fix was to wrap the Time::y2038 calls in code that duplicates
Time::Local's year handling.

Thanks to Slaven Rezic for finding this, and for having smokers
configured with modules like Time::y2038 (which is optional for this
package) so that he _could_ find it. Kudos!