Fix test failures when local zone not determined.
This was a single test in t/format_template.t that attempted to use
DateTime::Calendar::Christian to format March 15, 44 BC. The problem is
referred from DateTime::TimeZone::Local, which can not determine the
local zone under Haiku, and dies.

Thanks to Serguei Trouchelle for the tests that exposed this.