Implement interactive method else().
This must be in the scope of an if ... then begin, and there can only be
one.

Doing this involved fixing a bug that caused nested if() calls to be
handled incorrectly.

As a side implementation, the unsupported dump() method got cleaned up,
and 'dump frame' and 'dump tokens' added.