::Warner no longer tries to format Template::Exception objects.
Instead __message() (in ::Locale) re-creates the object with punctuation
and new line if it needs to.

The problem this fixes is that the macro code flattened
Template::Exception objects and then re-threw them. When they were
caught they were just strings, and ::Warner was looking at the strings,
seeing that they contained template interpolations, and trying to
execute those. This failed because it was just a template fragment, and
the resultant message was undef.