Remove macro subcommand defaults unless in level1 mode.
The problem was I mistyped 'macro define ...' as 'macro defined ...' in
my init file, and got a macro named 'defined'. Didn't realize it for a
long time because I didn't use that one much. And then didn't remember
the subcommand defaulting mechanism and so couldn't figure out why I
didn't get an error. So it's gone now.