Refactor pass event handling.
Pass events are now sub-templates named <main>:<thing>:<kind>, where
<main> is the main template name (e.g. 'pass'), <thing> is the thing
handled ('event' being the only value for now), and <kind> being the
kind of thing, which for events is (e.g.) 'rise'. Sub-template
<main>:<thing> handles things without their own <kind> template.

Note that as a part of this the heading generated by the 'pass' template
was tweaked for convenience of coding. Specifically, the 'Time' heading
is now left-justified rather than right-justified.